move tests into test_group_layer_widget
